refactor: migrate integration to astro-integration-kit
