+++ package.json
@@ -44,18 +44,20 @@
   ],
   "dependencies": {
     "@playwright/browser-chromium": "^1.54.2",
+    "astro-integration-kit": "^0.18.0",
     "get-port": "^7.1.0",
     "micro": "^10.0.1",
     "playwright": "^1.54.2",
     "serve-handler": "^6.1.6"
   },
   "devDependencies": {
-    "@types/react": "^19.1.9",
     "@types/serve-handler": "^6.1.4",
-    "astro": "^5.12.0",
     "prettier": "^3.6.2",
     "typescript": "^5.9.2",
     "xo": "^1.2.1"
+  },
+  "peerDependencies": {
+    "astro": "^5.7.0"
   },
   "xo": {
     "space": true,
